kotlin migration

+++ services/gateway/gateway-api/src/main/java/com/example/gateway/api/converter/TradeSummaryConverter.java
@@ -3,6 +3,7 @@ package com.example.gateway.api.converter;
 import com.example.gateway.api.spec.model.TradeSummaryGWModel;
 import com.example.tradelog.api.spec.model.TradeSummaryModel;
 
+import java.math.BigDecimal;
 import java.util.function.Function;
 
 public class TradeSummaryConverter {
@@ -10,7 +11,7 @@ public class TradeSummaryConverter {
     public static Function<TradeSummaryModel, TradeSummaryGWModel> toTradeSummaryGWModel = model ->
             TradeSummaryGWModel.builder()
                     .withSymbol(model.getSymbol())
-                    .withTotal(model.getTotal())
+                    .withTotal(BigDecimal.valueOf(model.getTotal()))
                     .withTrades(model.getTrades())
                     .build();
 }
